app/API/requirements.txt
fastapi
uvicorn
selenium
webdriver-manager
requests
beautifulsoup4
python-dotenv
openai
logging
pydantic